requirements.txt
fastapi
uvicorn
google-genai
fastapi-clerk-auth